# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://gestion-hotel.test/"
        }
    ],
    "paths": {
        "/api/habitacion": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener lista de habitaciones paginada",
                "operationId": "6731eddd075484fc40615891444029a8",
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Crear una nueva habitación",
                "description": "Crea una nueva habitación con los datos proporcionados",
                "operationId": "c88685cb8b9da97c1d2c634c57a5efba",
                "requestBody": {
                    "description": "Datos necesarios para crear una habitación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHabitacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/habitacion/{habitacion}": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener detalles de una habitación",
                "description": "Devuelve los detalles de una habitación específica",
                "operationId": "44dde2f8e2652fa501cfb8bbbc5a19f2",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la habitación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Actualizar una habitación",
                "description": "Actualiza los datos de una habitación específica",
                "operationId": "6736f90cf2de9d6174634d4c3724bca3",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para actualizar una habitación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHabitacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Habitación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Eliminar una habitación",
                "description": "Elimina una habitación específica por su ID",
                "operationId": "8dfea0e0a351fecff772c6187e218b73",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "ok"
                            }
                        }
                    }
                }
            }
        },
        "/api/habitacion/{habitacion}/hotel": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener el hotel asociado a una habitación",
                "description": "Devuelve el hotel asociado a una habitación específica.",
                "operationId": "2d23c06f26e7a1389a49686377163a02",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del hotel asociado a la habitación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Habitación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Habitación no encontrada"
                            }
                        }
                    }
                }
            }
        },
        "/api/habitacion/{habitacion}/huespedes": {
            "get": {
                "tags": [
                    "Habitación"
                ],
                "summary": "Obtener los huéspedes asociados a una habitación",
                "description": "Devuelve la lista de huéspedes asociados a una habitación específica.",
                "operationId": "65c05fef8d8b6f22eeeac0cdcf2cf9a2",
                "parameters": [
                    {
                        "name": "habitacion",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes asociados a la habitación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Huesped"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Habitación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Habitación no encontrada"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Hotel": {
                "title": "Hotel",
                "description": "Modelo de un hotel",
                "required": [
                    "nombre",
                    "direccion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del hotel",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del hotel",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del hotel",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono de contacto del hotel",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del hotel",
                        "type": "string",
                        "format": "email"
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Habitacion": {
                "title": "Habitación",
                "description": "Modelo de una habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la habitación",
                        "type": "integer"
                    },
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string"
                    },
                    "tipo": {
                        "description": "Tipo de habitación (ej. individual, doble)",
                        "type": "string"
                    },
                    "precioNoche": {
                        "description": "Precio por noche de la habitación",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece la habitación",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreHabitacionRequest": {
                "title": "Store Habitación Request",
                "description": "Datos requeridos para crear una nueva habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string",
                        "maxLength": 10
                    },
                    "tipo": {
                        "description": "Tipo de habitación",
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 5
                    },
                    "precioNoche": {
                        "description": "Precio por noche",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PutHabitacionRequest": {
                "title": "Put Habitación Request",
                "description": "Datos requeridos para actualizar una habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string",
                        "maxLength": 10
                    },
                    "tipo": {
                        "description": "Tipo de habitación",
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 5
                    },
                    "precioNoche": {
                        "description": "Precio por noche",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Huesped": {
                "title": "Huesped",
                "description": "Modelo de un huésped",
                "required": [
                    "nombre",
                    "apellido",
                    "dniPasaporte"
                ],
                "properties": {
                    "id": {
                        "description": "ID del huésped",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del huésped",
                        "type": "string"
                    },
                    "apellido": {
                        "description": "Apellido del huésped",
                        "type": "string"
                    },
                    "dniPasaporte": {
                        "description": "DNI o pasaporte del huésped",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}